feat: revamp homepage layout and navigation, update contact email
- Refactored the homepage (`src/app/home/page.tsx`) to use a new layout with a hero section, feature highlights, and financial solutions. Replaced old sections with new components.
- Updated the contact email in `ContactSection.tsx` to be a mailto link for easier user interaction.
- Moved the original homepage content to `src/app/page.tsx`.
